feat: clean up query module and grammar
